everything seems to work as before, although example kwics for parallel was broken before
