+++ app/index.js
@@ -97,8 +97,7 @@ require("./scripts/cqp_parser/CQPParser.js")
 require("./scripts/cqp_parser/cqp.js")
 require("./scripts/util.js")
 require("./scripts/pie-widget.js")
-require("./scripts/results.js")
-require("./scripts/trend_diagram/trend_results.js")
+require("./scripts/results/results.js")
 require("./scripts/model.js")
 require("./scripts/widgets.js")
 require("./scripts/main.js")
